+++ content/collection_popup.js
@@ -1,8 +1,25 @@
-//Get all different structures
-const main_body =   document.getElementById("main_body");
-const container =   document.getElementById("kanji_grid");
-const middle =      document.getElementById("middle");
+//Global elements-------------------------------------------
+const main_body =   document.getElementById("main-body");
 const tab_bar =     document.getElementById("tab-bar");
+
+//Collection tab elements-----------------------------------
+const collection_container =   document.getElementById("kanji-grid");
+
+//Account tab elements--------------------------------------
+const account_login =           document.getElementById("account-screen-login");
+const account_main =            document.getElementById("account-screen-main");
+const account_error =           document.getElementById("login-auth-error");
+const account_login_btn =       document.getElementById("login-btn-login");
+const account_register_btn =    document.getElementById("login-btn-register");
+const account_email_field =     document.getElementById("login-email");
+const account_pass_field =      document.getElementById("login-password");
+const account_logout_btn =      document.getElementById("account-btn-logout");
+const account_delete_btn =      document.getElementById("account-btn-delete");
+const account_change_pass_btn = document.getElementById("account-btn-change-pass");
+
+//Settings tab elements-------------------------------------
+const settings_palette_dark =   document.getElementById("palette-dark");
+const settings_palette_light =   document.getElementById("palette-light");
 
 //Create floating kanji card but hide it by default (like with littlepanel)
 const kanji_card_data = document.createElement("div");
@@ -13,6 +30,8 @@ document.body.appendChild(kanji_card_data);
 //Attach events to document
 create_events();
 
+
+// Collection Tab------------------------------------------------------------------------------------
 async function render_collection()
 {
     //Middle Section HTML----------------------------------------
@@ -38,12 +57,12 @@ async function render_collection()
         if(e.key === "Enter")
         {
             //Hide data card and show loading screen before request
-            container.style.display = "flex";
-            container.innerHTML = `<p class="collection-text-warnings">Loading...</p>`;
+            collection_container.style.display = "flex";
+            collection_container.innerHTML = `<p class="collection-text-warnings">Loading...</p>`;
             kanji_card_data.style.display = "none";
 
             //Set display mode to grid before displaying kanjis.
-            container.style.display = "grid";
+            collection_container.style.display = "grid";
 
             //Early return if there is no connection with server. check_connection() in api.js
             if(!check_connection())
@@ -85,17 +104,16 @@ async function render_collection()
             }
 
             //render grid with only matching results
-            container.innerHTML = get_grid_HTML(matches);
+            collection_container.innerHTML = get_grid_HTML(matches);
         }
     });
 }
 
-//Helpers----------------------------------------------------------------------
 //This function injects a text inside the middle section.
 function display_big_message(message)
 {
-    container.style.display = "flex";
-    container.innerHTML = `<p id="no_results" class="collection-text-warnings">${message}</p>`;
+    collection_container.style.display = "flex";
+    collection_container.innerHTML = `<p id="no_results" class="collection-text-warnings">${message}</p>`;
 }
 
 
@@ -150,7 +168,7 @@ async function render_card_data(e, kanji)
         await remove_kanji(btn.dataset.char);
 
         //Remove card from the grid
-        const card = container.querySelector(`[data-char="${btn.dataset.char}"]`);
+        const card = collection_container.querySelector(`[data-char="${btn.dataset.char}"]`);
         if(card) card.remove(); 
 
         //Update db size UI
@@ -175,7 +193,7 @@ async function render_grid(sorting = 'd')
 
     //if collection is empty, display message and return
     if (entries.length == 0) {
-        container.innerHTML = `
+        collection_container.innerHTML = `
             <p style="size: 1.5rem; font-weight: bold;">
                 No kanji saved yet.
             </p>`;
@@ -200,7 +218,7 @@ async function render_grid(sorting = 'd')
             break;
     }
     //Render kanji grid
-    container.innerHTML = get_grid_HTML(entries);
+    collection_container.innerHTML = get_grid_HTML(entries);
 }
 
 function get_grid_HTML(list)
@@ -221,9 +239,137 @@ function get_grid_HTML(list)
     return HTML;
 }
 
+// Account Tab --------------------------------------------------------------------------------------
+
+function account_show_screen(screen)
+{
+    account_login.style.display = 
+        screen === "login" ? "flex" : "none";
+    
+    account_main.style.display = 
+        screen === "main" ? "flex" : "none";
+}
+
+function account_show_error(message)
+{
+    account_error.textContent = message;
+    account_error.style.display = "block";
+}
+
+async function api_auth(endpoint, email, password)
+{
+    //Send a request to the API to get an auth token.
+    try
+    {
+        const res = await fetch(`${API_BASE}/auth/${endpoint}`, {
+            method: "POST",
+            headers: {"Content-Type": "application/json"},
+            body: JSON.stringify({email, password}),
+        });
+        return res;
+    }
+    catch(e)
+    {
+        return null;
+    }
+}
+
+async function initialize_account_tab()
+{
+    //Attempt to get auth token.
+    const token = await get_token();
+
+    //If no token is found, show login screen.
+    if(!token)
+    {
+        account_show_screen("login");
+        return;
+    }
+
+    //Check connection before showing main account screen.
+    //If connection with server is not possible, show login screen.
+    const connection_ok = check_connection();   //in shared/api.js
+    if(connection_ok)
+    {
+        account_show_screen("main");
+    }
+    else
+    {
+        account_show_screen("login");
+        account_show_error("Could not reach server.");
+    }
+}
+
+account_login_btn.addEventListener("click", async() => {
+    const email = account_email_field.value.trim();
+    const password = account_pass_field.value;
+    //null check on both fields is done on the back end.
+
+    const res = await api_auth("login", email, password);
+    if(!res) return account_show_error("Could not reach server.");
+
+    const data = await res.json();
+    //If response status code is outside the 200-299 range, something went wrong.
+    //In that case, api attaches an "error" field to the response's json. Display it.
+    if (!res.ok) return account_show_error(data.error || "Login failed.");
+
+    //If successful, get token from respoonse.
+    //Then display main 
+    await set_token(data.token);
+    account_show_screen("main");
+    display_tab("collection");
+});
+
+account_register_btn.addEventListener("click", async() => {
+    const email = account_email_field.value.trim();
+    const password = account_pass_field.value;
+
+    const res = await api_auth("register", email, password);
+    if(!res) return account_show_error("Could not reach server.");
+
+    const data = await res.json();
+    if(!res.ok) return account_show_error(data.error || "Registration failed.");
+
+    await set_token(data.token);
+    display_tab("collection");
+});
+
+
+account_logout_btn.addEventListener("click", async() => {
+
+    //clear auth token and show login screen.
+    await clear_token();
+    account_show_screen("login");
+});
+
+
+// Settings Tab -------------------------------------------------------------------------------------
+
+settings_palette_dark.addEventListener("click", () => {
+    choose_palette("default");
+});
+
+settings_palette_light.addEventListener("click", () => {
+    choose_palette("default-light");
+})
+
+// Global Methods------------------------------------------------------------------------------------
+//Acceptable values: "account", "settings", "collection"
+function display_tab(tab)
+{
+    document.querySelectorAll(".tab-panel").forEach(p => p.style.display = "none");
+    document.getElementById(`tab-${tab}`).style.display = "block";
+
+    //deselect all buttons
+    document.querySelectorAll(".btn-tab-active").forEach(b => b.className = "btn-tab");
+
+    //select "tab"'s button 
+    document.getElementById(`btn-tab-${tab}`).className = "btn-tab-active";
+}
+
 function create_events()
 { 
-    container.addEventListener("click", async (e) => { 
+    collection_container.addEventListener("click", async (e) => { 
         //get card closest to the cursor
         const card = e.target.closest(".kanji-card");
 
@@ -254,4 +400,6 @@ function create_events()
     });
 }
 
+
 render_collection();
+initialize_account_tab();